Refactor: 네이밍 통일성을 위한 패키지 이름 변경
- place.domain -> place.entity
